requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 
"livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0"
"livekit-plugins-noise-cancellation~=0.2"